requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
google-genai>=0.10.0
